Port factory.js tests to xulrunner.
