+++ angel-player/src/chrome/content/test/simpletests/factory/size_ngl_obj.js
@@ -1,0 +1,15 @@
+var buffer = require ( 'tenshi/angelic/buffer.js' );
+var misc = require ( 'tenshi/angelic/misc.js' );
+var factory = require ( 'tenshi/common/factory.js' );
+var url = require ( 'jetpack/sdk/url' );
+
+exports.run = function() {
+  var fact = factory.make ( );
+  fact.set_target_type ( 'ARM' );
+  fact.load_type_file ( url.toFilename ( 
+        'chrome://angel-player/content/common_defs/ngl_types.yaml' ) );
+
+  misc.assert ( fact.get_size ( 'ngl_obj' ) === 4 * 2,
+                'Size of ngl_obj should be 8 bytes.' );
+  return true;
+};
